requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing-extensions==4.12.2
tzdata==2024.1
werkzeug==3.0.3
zipp==3.19.2
